requirements.txt
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.24.3
opencv-python==4.8.1.78
packaging==25.0
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.10
python3-xlib==0.15
pytweening==1.2.0
